refactor(init)!: merge non-plugin config into a single init.lua file for ease of maintenance
