+++ init.lua
@@ -4,55 +4,336 @@
 ---      /\_______/\
 ---     /   o   o   \
 ---    (  ==  ^  ==  )
----     )           (
+---     )  [Ollie]  (
 ---    (             )
 ---    ( (  )   (  ) )
 ---   (__(__)___(__)__)
 ---  ___
 ---   | |_  _  _     o __
 ---   | | |(/_(_)\_/ | |||
+---  Oliver : https://www.asciiart.eu/animals/cats + Jonathan added a few layers of belly because Oliver is a chunky boi
 ---
---- Cat : https://www.asciiart.eu/animals/cats
----      My friend added a few layers of belly so it looks more like chunky my cat Oliver
---- Logo: $ figlet -f mini Theovim
----
---- Initialize all configuration files
 
---- safe_require()
---- Tries calling `require` for the given module.
---- If unsuccessful, prints the error message using `vim.notify()`
----
----@param module string name of the module to `require`
----@return unknown module from `pcall` if the call was successful, otherwise nil
-local function safe_require(module)
-  local status, loaded_module = pcall(require, module)
-  if status then
-    return loaded_module
+-- {{{ options
+-- Excludes default Neovim options:
+-- https://neovim.io/doc/user/vim_diff.html#nvim-defaults
+local opt          = vim.opt
+
+-- Tab
+opt.softtabstop    = 0    --> How many chracters the /cursor moves/ with <TAB> and <BS> -- 0 to disable
+opt.expandtab      = true --> Use space instead of tab
+opt.shiftwidth     = 2    --> Number of spaces to use for auto-indentation, <<, >>, etc.
+opt.shiftround     = true --> Make the indentation to a multiple of shiftwidth when using < or >
+
+-- Location in the buffer
+opt.number         = true
+opt.relativenumber = true
+opt.cursorline     = true
+opt.cursorlineopt  = "number" --> line, screenline, both (i.e., "number,line")
+opt.cursorcolumn   = true
+
+-- Search and replace
+opt.ignorecase     = true    --> Ignore case in search
+opt.smartcase      = true    --> /smartcase -> apply ignorecase | /sMartcase -> do not apply ignorecase
+opt.inccommand     = "split" --> show the substitution in a split window
+
+-- Split
+opt.splitright     = true --> Vertical split created right
+opt.splitbelow     = true --> Horizontal split created below
+
+-- UI
+opt.signcolumn     = "yes"          --> Render signcolumn always to prevent text shifting
+opt.scrolloff      = 7              --> Keep minimum x number of screen lines above and below the cursor
+opt.termguicolors  = true           --> Enables 24-bit RGB color in the TUI
+opt.showtabline    = 2              --> 0: never, 1: >= 2 tabs, 2: always
+opt.laststatus     = 2              --> 0: never, 1: >= 2 windows, 2: always, 3: always and have one global statusline
+
+-- Char rendering
+opt.list           = true --> Render special char in listchars
+opt.listchars      = { tab = "⇥ ", trail = "␣", nbsp = "⍽", leadmultispace = "┊ ", }
+opt.showbreak      = "↪" --> Render beginning of wrapped lines
+opt.breakindent    = true --> Wrapped line will have the same indentation level as the beginning of the line
+
+-- Spell
+opt.spell          = false    --> autocmd will enable spellcheck in Tex or markdown
+opt.spelllang      = { "en" }
+opt.spellsuggest   = "best,8" --> Suggest 8 words for spell suggestion
+opt.spelloptions   = "camel"  --> Consider CamelCase when checking spelling
+
+-- Fold
+opt.foldenable     = false                        --> Open all folds until I close them using zc/zC or update using zx
+opt.foldmethod     = "expr"                       --> Use `foldexpr` function for folding
+opt.foldexpr       = "nvim_treesitter#foldexpr()" --> Treesitter folding
+--foldlevel      = 2                            --> Ignore n - 1 level fold
+
+-- Update time
+opt.updatetime     = 250
+opt.timeoutlen     = 300
+
+-- Window size
+opt.winminwidth    = 3
+
+-- Others
+opt.mouse          = "a"
+opt.confirm        = true --> Confirm before exiting with unsaved bufffer(s)
+-- }}}
+
+
+-- {{{ keymaps
+local set = vim.keymap.set
+
+-- Space as the leader
+set({ "n", "v" }, "<Space>", "<Nop>")
+vim.g.mapleader = " "
+vim.g.maplocalleader = " "
+
+-- Default overrides
+set("n", "<ESC>", "<CMD>nohlsearch<CR>")
+set("t", "<ESC><ESC>", "<C-\\><C-n>", { desc = "Exit terminal mode" })
+
+set("n", "k", "v:count == 0 ? 'gk' : 'k'", { expr = true })
+set("n", "j", "v:count == 0 ? 'gj' : 'j'", { expr = true })
+
+set("n", "n", "nzz")
+set("n", "N", "Nzz")
+set("n", "<C-u>", "<C-u>zz")
+set("n", "<C-d>", "<C-d>zz")
+
+set("n", "<C-w>+", "<C-w>+<CMD>call feedkeys('<C-w>')<CR>", { desc = "Increase the window height (press + to repeat)" })
+set("n", "<C-w>-", "<C-w>-<CMD>call feedkeys('<C-w>')<CR>", { desc = "Decrease the window height (press - to repeat)" })
+set("n", "<C-w>>", "<C-w>><CMD>call feedkeys('<C-w>')<CR>", { desc = "Increase the window width (press > to repeat)" })
+set("n", "<C-w><", "<C-w><<CMD>call feedkeys('<C-w>')<CR>", { desc = "Decrease the window width (press < to repeat)" })
+
+-- Custom keymaps
+set("i", "jk", "<ESC>", { desc = "Better ESC" })
+set("i", "<C-s>", "<C-g>u<ESC>[s1z=`]a<C-g>u", { desc = "Fix nearest [S]pelling error and put the cursor back" })
+
+-- Copy and paste
+set({ "n", "x" }, "<leader>a", "gg<S-v>G", { desc = "Select [A]ll" })
+set("x", "<leader>y", '"+y', { desc = "[Y]ank to the system clipboard (+)" })
+set("n",
+  "<leader>p",
+  ":echo '[Theovim] e.g.: `:normal \"3p` to paste the content of the register 3'<CR>" .. ':reg<CR>:normal "',
+  { silent = true, desc = "[P]aste from one of the registers" })
+set("x",
+  "<leader>p",
+  '"_dP', --> [d]elete the selection and send content to _ void reg then [P]aste (b4 cursor unlike small p)
+  { desc = "[P]aste the current selection without overriding the register" })
+
+-- Buffer
+set("n", "[b", "<CMD>bprevious<CR>", { desc = "Go to previous [B]uffer" })
+set("n", "]b", "<CMD>bnext<CR>", { desc = "Go to next [B]uffer" })
+set("n",
+  "<leader>b",
+  ":echo '[Theovim] Choose a buffer'<CR>" .. ":ls<CR>" .. ":b<SPACE>",
+  { silent = true, desc = "Open [B]uffer list" })
+set("n",
+  "<leader>k",
+  ":echo '[Theovim] Choose a buf to delete (blank to choose curr)'<CR>" .. ":ls<CR>" .. ":bdelete<SPACE>",
+  { silent = true, desc = "[K]ill a buffer" })
+
+-- Terminal
+-- https://medium.com/@egzvor/vim-splits-cheat-sheet-2bf84fc99962
+--           | :top sp |
+-- |:top vs| |:abo| cu | |:bot vs |
+-- |       | |:bel| rr | |        |
+--           | :bot sp |
+-- botright == bot
+set("n",
+  "<leader>tb",
+  function()
+    vim.cmd("botright " .. math.ceil(vim.fn.winheight(0) * (1 / 3)) .. "sp | term")
+  end,
+  { desc = "Launch a [t]erminal in the [B]ottom" })
+
+set("n",
+  "<leader>tr",
+  function()
+    vim.cmd("bot " .. math.ceil(vim.fn.winwidth(0) * 0.3) .. "vs | term")
+  end,
+  { desc = "Launch a [t]erminal to the [R]ight" })
+
+set("n",
+  "<leader>tt",
+  function()
+    vim.cmd("tabnew | term")
+  end,
+  { desc = "Launch a [t]erminal in a new [T]ab" })
+
+--- Move to a window (one of hjkl) or create a split if a window does not exist in the direction.
+--- Lua translation of:
+--- https://www.reddit.com/r/vim/comments/166a3ij/comment/jyivcnl/?utm_source=share&utm_medium=web2x&context=3
+--- Usage: vim.keymap("n", "<C-h>", function() move_or_create_win("h") end, {})
+--
+---@param key string One of h, j, k, l, a direction to move or create a split
+local function smarter_win_nav(key)
+  local fn = vim.fn
+  local curr_win = fn.winnr()
+  vim.cmd("wincmd " .. key)        --> attempt to move
+
+  if (curr_win == fn.winnr()) then --> didn't move, so create a split
+    if key == "h" or key == "l" then
+      vim.cmd("wincmd v")
+    else
+      vim.cmd("wincmd s")
+    end
+
+    vim.cmd("wincmd " .. key) --> move again
   end
-  vim.notify("Error loading the module: " .. module)
-  return nil
 end
 
--- Loads core config modules
-safe_require("config.opt")
-safe_require("config.keymap")
-safe_require("config.command")
-safe_require("config.autocmd")
-safe_require("config.netrw")
+set("n", "<C-h>", function() smarter_win_nav("h") end,
+  { desc = "Move focus to the left window or create a horizontal split" })
+set("n", "<C-j>", function() smarter_win_nav("j") end,
+  { desc = "Move focus to the lower window or create a vertical split" })
+set("n", "<C-k>", function() smarter_win_nav("k") end,
+  { desc = "Move focus to the upper window or create a vertical split" })
+set("n", "<C-l>", function() smarter_win_nav("l") end,
+  { desc = "Move focus to the right window or create a horizontal split" })
+-- }}}
 
--- Loads a built-in diagnostic config
-safe_require("config.diagnostic")
 
--- Loads a config for Lazy.nvim
--- lazy.lua will then
--- 1. Combine Lua modules (returned tables) in `lua/plugins` directory
--- 2. Calls `config` or `setup(opts)` to loads plugins
-safe_require("config.lazy")
+-- {{{ commands
+--- Trims trailing whitespace
+--- \s: white space char, \+ :one or more, $: end of the line, e: suppresses warning when no match found, c: confirm
+local function trim_whitespace()
+  local win_save = vim.fn.winsaveview()
+  vim.cmd([[keeppatterns %s/\s\+$//ec]])
+  vim.fn.winrestview(win_save)
+end
+vim.api.nvim_create_user_command("TrimWhitespace", trim_whitespace, { nargs = 0 })
 
--- Calls the init.lua of the Theovim built-in UI module package
-safe_require("ui")
+-- Changes current working directory
+vim.api.nvim_create_user_command("CD", ":lcd %:h", { nargs = 0 })
+-- }}}
 
--- Configures Neovide
+
+-- {{{ autocmds
+-- Highlight on yank
+vim.api.nvim_create_autocmd("TextYankPost", {
+  group = vim.api.nvim_create_augroup("TheovimYankHighlight", { clear = true }),
+  pattern = "*",
+  callback = function() vim.highlight.on_yank() end,
+  desc = "Highlight yanked text",
+})
+
+-- Switch to insert mode when terminal is open
+vim.api.nvim_create_autocmd({ "TermOpen", "BufEnter" }, {
+  -- TermOpen: for when terminal is opened for the first time
+  -- BufEnter: when you navigate to an existing terminal buffer
+  group = vim.api.nvim_create_augroup("TheovimTerminal", { clear = true }),
+  pattern = "term://*", --> only applicable for "BufEnter", an ignored Lua table key when evaluating TermOpen
+  callback = function() vim.cmd("startinsert") end
+})
+
+-- Update indentation guide dynamically
+local update_leadmultispace_group = vim.api.nvim_create_augroup("UpdateLeadmultispace", { clear = true })
+
+--- Dynamically adjust `leadmultispace` in `listchars` (buffer level) based on `shiftwidth`
+local function update_leadmultispace()
+  local lead = "┊"
+  for _ = 1, vim.bo.shiftwidth - 1 do
+    lead = lead .. " "
+  end
+  vim.opt_local.listchars:append({ leadmultispace = lead })
+end
+
+-- When `shiftwidth` was manually changed
+vim.api.nvim_create_autocmd("OptionSet", {
+  group = update_leadmultispace_group,
+  pattern = { "shiftwidth", "filetype" },
+  callback = update_leadmultispace,
+})
+
+-- When shiftwidth was changed by ftplugin
+vim.api.nvim_create_autocmd("BufEnter", {
+  group = update_leadmultispace_group,
+  pattern = "*",
+  callback = update_leadmultispace,
+  once = true,
+})
+-- }}}
+
+
+-- {{{ netrw
+vim.g.netrw_banner = 0
+vim.g.netrw_liststyle = 0    --> 0: Simple, 1: Detailed, 2: Thick, 3: Tree
+vim.g.netrw_browse_split = 3 --> Open file in 0: Reuse the same win, 1: Horizontal split, 2: Vertical split, 3: New tab
+vim.g.netrw_winsize = 25     --> seems to be in percentage
+
+vim.g.is_netrw_open = false
+--- Toggles netrw
+--- Requires vim.g.is_netrw_open global variable to function properly
+local function toggle_netrw()
+  local fn = vim.fn
+  if vim.g.is_netrw_open then
+    for i = 1, fn.bufnr("$") do
+      if fn.getbufvar(i, "&filetype") == "netrw" then
+        vim.cmd("bwipeout " .. i)
+      end
+    end
+    vim.g.is_netrw_open = false
+  else
+    vim.cmd("Lex")
+    vim.g.is_netrw_open = true
+  end
+end
+vim.keymap.set("n", "<leader>n", toggle_netrw,
+  { silent = true, noremap = true, desc = "Toggle [N]etrw" })
+-- }}}
+
+
+-- {{{ diagnostic
+-- Diagnostic appearance
+vim.diagnostic.config({
+  float = {
+    border = "rounded",
+    format = function(diagnostic)
+      -- "ERROR (line n): message"
+      return string.format("%s (line %i): %s",
+        vim.diagnostic.severity[diagnostic.severity],
+        diagnostic.lnum + 1,
+        diagnostic.message)
+    end
+  },
+  update_in_insert = false,
+})
+
+-- Keymaps
+vim.keymap.set("n", "<leader>q", vim.diagnostic.setloclist, { desc = "Open diagnostic [Q]uickfix list" })
+-- }}}
+
+
+-- {{{ lazy.nvim
+-- Bootstrap lazy.nvim
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not (vim.uv or vim.loop).fs_stat(lazypath) then
+  local lazyrepo = "https://github.com/folke/lazy.nvim.git"
+  local out = vim.fn.system({ "git", "clone", "--filter=blob:none", "--branch=stable", lazyrepo, lazypath })
+  if vim.v.shell_error ~= 0 then
+    vim.api.nvim_echo({
+      { "Failed to clone lazy.nvim:\n", "ErrorMsg" },
+      { out,                            "WarningMsg" },
+      --{ "\nPress any key to exit..." },
+    }, true, {})
+    --vim.fn.getchar()
+    --os.exit(1)
+  end
+end
+vim.opt.rtp:prepend(lazypath)
+
+-- Either pass the table (containing the plugin info) or string (a directory containing Lua modules)
+-- Following will import all Lua files in lua/plugins
+require("lazy").setup("plugins")
+-- }}}
+
+-- {{{ UI module calls
+require("ui.statusline").setup()
+require("ui.tabline").setup()
+require("ui.dashboard").setup()
+-- }}}
+
+-- {{{ Neovide settings
 if vim.g.neovide then
   local padding = 10
   vim.g.neovide_padding_top = padding
@@ -67,3 +348,6 @@ if vim.g.neovide then
   vim.g.neovide_floating_blur_amount_x = 2.0
   vim.g.neovide_floating_blur_amount_y = 2.0
 end
+-- }}}
+
+-- vim: foldmethod=marker
