--- lua/config/netrw.lua
@@ -1,35 +1,0 @@
---- config/netrw.lua
----
---- $ figlet -f small theovim
----  _   _                _
---- | |_| |_  ___ _____ _(_)_ __
---- |  _| ' \/ -_) _ \ V / | '  \
----  \__|_||_\___\___/\_/|_|_|_|_|
----
---- netrw file browser settings
----
-
-vim.g.netrw_banner = 0
-vim.g.netrw_liststyle = 0    --> 0: Simple, 1: Detailed, 2: Thick, 3: Tree
-vim.g.netrw_browse_split = 3 --> Open file in 0: Reuse the same win, 1: Horizontal split, 2: Vertical split, 3: New tab
-vim.g.netrw_winsize = 25     --> seems to be in percentage
-
-vim.g.is_netrw_open = false
---- Toggles netrw
---- Requires vim.g.is_netrw_open global variable to function properly
-local function toggle_netrw()
-  local fn = vim.fn
-  if vim.g.is_netrw_open then
-    for i = 1, fn.bufnr("$") do
-      if fn.getbufvar(i, "&filetype") == "netrw" then
-        vim.cmd("bwipeout " .. i)
-      end
-    end
-    vim.g.is_netrw_open = false
-  else
-    vim.cmd("Lex")
-    vim.g.is_netrw_open = true
-  end
-end
-vim.keymap.set("n", "<leader>n", toggle_netrw,
-  { silent = true, noremap = true, desc = "Toggle [N]etrw" })
